+++ src/main/java/kz/nmbet/betradar/dao/repository/GlOutrightOddEntityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlOutrightOddEntityRepository extends JpaRepository<GlOutrightOddEntity, Long> {
+public interface GlOutrightOddEntityRepository extends JpaRepository<GlOutrightOddEntity, Integer> {
 
 	
 }
