+++ src/main/java/kz/nmbet/betradar/dao/repository/GlPlayerEntityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlPlayerEntityRepository extends JpaRepository<GlPlayerEntity, Long> {
+public interface GlPlayerEntityRepository extends JpaRepository<GlPlayerEntity, Integer> {
 
 	
 }
