+++ src/main/java/kz/nmbet/betradar/dao/repository/GlSportEntityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlSportEntityRepository extends JpaRepository<GlSportEntity, Long> {
+public interface GlSportEntityRepository extends JpaRepository<GlSportEntity, Integer> {
 
 	
 }
