+++ src/main/java/kz/nmbet/betradar/dao/repository/GlTournamentEntityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlTournamentEntityRepository extends JpaRepository<GlTournamentEntity, Long> {
+public interface GlTournamentEntityRepository extends JpaRepository<GlTournamentEntity, Integer> {
 
 	
 }
