+++ src/main/java/kz/nmbet/betradar/dao/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<GlUser, Long> {
+public interface UserRepository extends JpaRepository<GlUser, Integer> {
 
 	GlUser findByEmail(String email);
 }
